Fix path name

+++ src/main/java/com/xhdh/xhdh/dto/VoteRequest.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 @Setter
 public class VoteRequest {
     private long userId;
-    private int universityId;
-    private int tagId;
+    private long universityId;
+    private long tagId;
     private long matchId;
 }
